+++ packages/apple_vision_animal_pose/lib/src/animal_pose_controller.dart
@@ -16,7 +16,7 @@ class AppleVisionAnimalPoseController {
   /// 
   /// [imageSize] as Size is the size of the image that is being processed
   /// 
-  /// [orientation]
+  /// [orientation] The orientation of the image
   Future<List<AnimalPoseData>?> processImage(Uint8List image, Size imageSize,[ImageOrientation orientation = ImageOrientation.downMirrored]) async{
     try {
       final Map<String, dynamic>? result = await _methodChannel.invokeMapMethod<String, dynamic>(