+++ packages/apple_vision_face/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.0.3
+
+* Added image orientation
+
 ## 0.0.2
 
 * Updates to example files, and fixed point offset