+++ packages/apple_vision_selfie/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.0.2
+
+* Added image orientation
+
 ## 0.0.1
 
 * Initial release